requirements.txt
certifi==2021.5.30
discord.py==1.7.3
ffmpeg==1.4
ffmpeg_python==0.2.0
gspread==4.0.1
oauth2client==4.1.3
pythonping==1.1.0
youtube_dl==2021.6.6
